API and Requirement changes
 - Introduce new calibration backend
 - Restructure Matrix class API:
    * Remove ._info structure and merge with shadowed properties
    * Introduce public and readonly declarations to eliminate property use
    * Remove structures holding calibration and performance data
 - Restructure profiling interface which is now based on transform calls.
   This allows easier extension in the future
 - Restructure MatrixCalibration class to reflect the new flexibility
 - Increase the requirement on cython. Due to auto_pickle support the version
   requirement was increased to >= 0.26. However due to intermediate bugfixes
   using the latest version is strongly recommenden
 - Introduce MatrixCallProfile (and document it) which now acts as central
   point for runtime performance evaluations during the processing tree
 - fastmat matrices are now compatible with the pickle protocol and pickling
   them tastes nice

+++ fastmat/LowRank.pyx
@@ -189,24 +189,26 @@ cdef class LowRank(Product):
             BENCH.COMMON: {
                 BENCH.FUNC_SIZE : (lambda c: 10 * c + 1),
                 BENCH.FUNC_GEN  : (lambda c: LowRank(
-                    arrTestDist((c + 1, ), dtype=np.float, center=2),
-                    arrTestDist((10 * c + 1, c + 1), dtype=np.float, center=2),
-                    arrTestDist((10 * c + 1, c + 1), dtype=np.float, center=2)))
+                    arrTestDist((c + 1, ), dtype=np.float32, center=2),
+                    arrTestDist((10 * c + 1, c + 1),
+                                dtype=np.float32, center=2),
+                    arrTestDist((10 * c + 1, c + 1),
+                                dtype=np.float32, center=2)))
             },
             BENCH.FORWARD: {},
             BENCH.OVERHEAD: {
                 BENCH.FUNC_SIZE : (lambda c: 2 ** c),
                 BENCH.FUNC_GEN  : (lambda c: LowRank(
-                    arrTestDist((c, ), dtype=np.float),
-                    arrTestDist((2 ** c, c), dtype=np.float),
-                    arrTestDist((2 ** c, c), dtype=np.float)))
+                    arrTestDist((c, ), dtype=np.float32),
+                    arrTestDist((2 ** c, c), dtype=np.float32),
+                    arrTestDist((2 ** c, c), dtype=np.float32)))
             },
             BENCH.DTYPES: {
                 BENCH.FUNC_SIZE : (lambda c: 2 ** c),
                 BENCH.FUNC_GEN  : (lambda c, datatype: LowRank(
-                    arrTestDist((c, ), dtype=np.float),
-                    arrTestDist((2 ** c, c), dtype=np.float),
-                    arrTestDist((2 ** c, c), dtype=np.float)))
+                    arrTestDist((c, ), dtype=np.float32),
+                    arrTestDist((2 ** c, c), dtype=np.float32),
+                    arrTestDist((2 ** c, c), dtype=np.float32)))
             }
         }
